pyproject.toml
[project]
name = "simplesocialnetwork"
version = "0.1.0"
description = ""
authors = [
    {name = "bonafides",email = "alexsandr.ejik@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
